+++ R/backend-redshift.R
@@ -171,7 +171,7 @@ sql_paste_redshift <- function(sep) {
 # https://docs.aws.amazon.com/redshift/latest/dg/r_EXPLAIN.html
 #' @export
 sql_query_explain.Redshift <- function(con, sql, ...) {
-  glue_sql2(con, "EXPLAIN {sql}")
+  sql_glue2(con, "EXPLAIN {sql}")
 }
 
 #' @export