+++ R/backend-sqlite.R
@@ -39,7 +39,7 @@ db_connection_describe.SQLiteConnection <- function(con, ...) {
 
 #' @export
 sql_query_explain.SQLiteConnection <- function(con, sql, ...) {
-  glue_sql2(con, "EXPLAIN QUERY PLAN {sql}")
+  sql_glue2(con, "EXPLAIN QUERY PLAN {sql}")
 }
 
 #' @export
@@ -132,7 +132,7 @@ sql_escape_logical.SQLiteConnection <- function(con, x) {
 #' @export
 sql_expr_matches.SQLiteConnection <- function(con, x, y, ...) {
   # https://sqlite.org/lang_expr.html#isisnot
-  glue_sql2(con, "{x} IS {y}")
+  sql_glue2(con, "{x} IS {y}")
 }
 
 #' @export