+++ R/backend-teradata.R
@@ -234,5 +234,5 @@ teradata_as_date <- function(x) {
 #' @export
 sql_table_analyze.Teradata <- function(con, table, ...) {
   # https://www.tutorialspoint.com/teradata/teradata_statistics.htm
-  glue_sql2(con, "COLLECT STATISTICS {.tbl table}")
+  sql_glue2(con, "COLLECT STATISTICS {.tbl table}")
 }